+++ src/dependency-manager/deprecated-spec/index.ts
@@ -1,6 +1,7 @@
-import { ComponentConfig } from '../..';
+
+import { ComponentConfig } from '../config/component-config';
 import { DependencyGraph } from '../graph';
-import DependencyManager from '../manager';
+import type DependencyManager from '../manager';
 
 export abstract class DeprecatedSpec {
   protected manager: DependencyManager;
